made so httpCode is always present in response

+++ classes/Repository/ProductRepository.php
@@ -107,7 +107,7 @@ class ProductRepository
             return 0;
         }
 
-        return (int) (count($products));
+        return count($products);
     }
 
     /**
